Add fetching background video depending on the weather condition functionality
