+++ src/index.js
@@ -1,13 +1,14 @@
-import "@babel/polyfill";
-import { call } from './logic';
+import '@babel/polyfill';
+import { getWeather } from './dom';
 import './style/style.css';
 
 const video = document.querySelector('#video');
+const city = document.querySelector('#city');
+const country = document.querySelector('#country');
+const tempurature = document.querySelector('#tempurature');
+const unit = document.querySelector('#tempurature');
 
 
 
-call(video);
-
-
-// pexels(video);
+getWeather(video);
